nohz: Exit RCU idle mode when calling preempt_schedule_irq()
When a CPU running tickless resumes userspace, it enters
RCU idle mode. But if we are preempted on irq exit through
preempt_schedule_irq(), we need to re-enable RCU in case
this function makes use of RCU read side critical section
and also for the next task to be scheduled.

TODO: merge into previous commit to avoid breaking bisect

Signed-off-by: Frederic Weisbecker <fweisbec@gmail.com>